Split Android AI runtime coordinators
