+++ apps/android/feature/ai/src/main/java/com/flashcardsopensourceapp/feature/ai/runtime/AiChatRuntime.kt
@@ -24,12 +24,12 @@ import com.flashcardsopensourceapp.feature.ai.runtime.conversation.canEditAiDraf
 import com.flashcardsopensourceapp.feature.ai.runtime.conversation.canManageAiDraftAttachments
 import com.flashcardsopensourceapp.feature.ai.runtime.conversation.canPrepareAiDraftInComposerPhase
 import com.flashcardsopensourceapp.feature.ai.runtime.conversation.shouldPrepareGuestAccess
-import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatBootstrapCoordinator
-import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatDictationCoordinator
-import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatLiveStreamCoordinator
-import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatRuntimeLifecycleCoordinator
-import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatSendCoordinator
-import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatSessionCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.bootstrap.AiChatBootstrapCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.dictation.AiChatDictationCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.lifecycle.AiChatRuntimeLifecycleCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.live.AiChatLiveStreamCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.send.AiChatSendCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.session.AiChatSessionCoordinator
 import com.flashcardsopensourceapp.feature.ai.runtime.errors.AiAlertState
 import com.flashcardsopensourceapp.feature.ai.runtime.observability.AiChatBreadcrumb
 import com.flashcardsopensourceapp.feature.ai.runtime.observability.AiChatWarning
