+++ apps/android/feature/ai/src/test/java/com/flashcardsopensourceapp/feature/ai/runtime/dictation/AiChatDictationCoordinatorTest.kt
@@ -3,8 +3,8 @@ package com.flashcardsopensourceapp.feature.ai.runtime
 import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiChatRuntimeState
 import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiComposerPhase
 import com.flashcardsopensourceapp.feature.ai.runtime.conversation.makeAiDraftState
-import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatDictationCoordinator
-import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatSessionCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.dictation.AiChatDictationCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.session.AiChatSessionCoordinator
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatDictationState
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatTranscriptionResult
 import com.flashcardsopensourceapp.data.local.model.ai.defaultAiChatServerConfig
